fix: Disable App Sandbox for Homebrew and SMC access
The app needs to:
1. Access /opt/homebrew/bin/brew to run Homebrew commands
2. Access SMC (System Management Controller) via IOKit for temperature reading

App Sandbox prevents both of these. Since this is a system utility app
that needs low-level hardware access, we disable the sandbox.

Note: This means the app cannot be distributed on the Mac App Store,
but it can be distributed via DMG/Homebrew cask (which is our target).